feat(#13) :: 컴포넌트 분리
추후 설정이나 다른 곳에서도 사용할 수 있을 가능성이 있어 컴포넌트 분리작업
